Make fewer things static for multithreading
Multithreading hates things that change "randomly". Therefore, we will
remove all static accesses of variables and make them explcit by adding
them to the parameter lists of the functions. Note that the way that
this is handled so far is by making the variables private instead of
public. This way, it is easier to make things per-thread.